# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ProcessMaker API",
        "description": "",
        "contact": {
            "email": "info@processmaker.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/1.0"
        }
    ],
    "paths": {
...
        "/screens": {
            "get": {
                "tags": [
                    "Screens"
                ],
                "summary": "Returns all screens that the user has access to",
                "operationId": "getScreens",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/filter"
                    },
                    {
                        "$ref": "#/components/parameters/order_by"
                    },
                    {
                        "$ref": "#/components/parameters/order_direction"
                    },
                    {
                        "$ref": "#/components/parameters/per_page"
                    },
                    {
                        "$ref": "#/components/parameters/include"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "list of screens",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "Get a list of Screens.",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/screens"
                                            }
                                        },
                                        "meta": {
                                            "description": "Get a list of Screens.",
                                            "type": "object",
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/metadata"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Screens"
                ],
                "summary": "Save a new screens",
                "operationId": "createScreens",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/screensEditable"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/screens"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/screens/{screens_id}": {
            "get": {
                "tags": [
                    "Screens"
                ],
                "summary": "Get single screens by ID",
                "operationId": "getScreensById",
                "parameters": [
                    {
                        "name": "screens_id",
                        "in": "path",
                        "description": "ID of screens to return",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully found the screen",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/screens"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Screens"
                ],
                "summary": "Update a screen",
                "operationId": "updateScreen",
                "parameters": [
                    {
                        "name": "screens_id",
                        "in": "path",
                        "description": "ID of screen to return",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/screensEditable"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "success"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Screens"
                ],
                "summary": "Delete a screen",
                "operationId": "deleteScreen",
                "parameters": [
                    {
                        "name": "screens_id",
                        "in": "path",
                        "description": "ID of screen to return",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "success"
                    }
                }
            }
        },
        "/screens/{screens_id}/duplicate": {
            "put": {
                "tags": [
                    "Screens"
                ],
                "summary": "duplicate a screen",
                "operationId": "duplicateScript",
                "parameters": [
                    {
                        "name": "screens_id",
                        "in": "path",
                        "description": "ID of screen to return",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/screensEditable"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/screens"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/screens/{screensId}/export": {
            "post": {
                "tags": [
                    "Screens"
                ],
                "summary": "Export a single screen by ID",
                "operationId": "exportScreen",
                "parameters": [
                    {
                        "name": "screensId",
                        "in": "path",
                        "description": "ID of screen to return",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully exported the screen",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/screenExported"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/screens/import": {
            "post": {
                "tags": [
                    "Screens"
                ],
                "summary": "Import a new screen",
                "operationId": "importScreen",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "file to upload",
                                        "type": "file",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/screens"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "screensEditable": {
                "properties": {
                    "title": {
                        "description": "Class Screen",
                        "type": "string"
                    },
                    "type": {
                        "description": "Class Screen",
                        "type": "string"
                    },
                    "description": {
                        "description": "Class Screen",
                        "type": "string"
                    },
                    "config": {
                        "description": "Class Screen",
                        "type": "string"
                    },
                    "computed": {
                        "description": "Class Screen",
                        "type": "string"
                    },
                    "custom_css": {
                        "description": "Class Screen",
                        "type": "string"
                    },
                    "screen_category_id": {
                        "description": "Class Screen",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "screens": {
                "properties": {
                    "id": {
                        "description": "Class Screen",
                        "type": "string",
                        "format": "id"
                    },
                    "created_at": {
                        "description": "Class Screen",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Class Screen",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/screensEditable"
                    }
                ]
            },
            "screenExported": {
                "properties": {
                    "url": {
                        "description": "Class Screen",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "metadata": {
                "properties": {
                    "filter": {
                        "type": "string"
                    },
                    "sort_by": {
                        "type": "string"
                    },
                    "sort_order": {
                        "type": "string",
                        "enum": [
                            "asc",
                            "desc"
                        ]
                    },
                    "count": {
                        "type": "integer"
                    },
                    "total_pages": {
                        "type": "integer"
                    },
                    "current_page": {
                        "type": "integer"
                    },
                    "form": {
                        "type": "integer"
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "path": {
                        "type": "string"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "to": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
            "filter": {
                "name": "filter",
                "in": "query",
                "description": "Filter results by string. Searches Name, Description, and Status. Status must match exactly. Others can be a substring.",
                "schema": {
                    "type": "string"
                }
            },
            "order_by": {
                "name": "order_by",
                "in": "query",
                "description": "Field to order results by",
                "schema": {
                    "type": "string"
                }
            },
...
            "order_direction": {
                "name": "order_direction",
                "in": "query",
                "schema": {
                    "type": "string",
                    "default": "asc",
                    "enum": [
                        "asc",
                        "desc"
                    ]
                }
            },
            "per_page": {
                "name": "per_page",
                "in": "query",
                "schema": {
                    "type": "integer",
                    "default": "10"
                }
            },
            "include": {
                "name": "include",
                "in": "query",
                "description": "Include data from related models in payload. Comma seperated list.",
                "schema": {
                    "type": "string",
                    "default": ""
                }
            },
...
        },
        "securitySchemes": {
            "pm_api_auth_code": {
                "type": "oauth2",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "/oauth/authorize",
                        "tokenUrl": "/oauth/token",
                        "scopes": {
                            "*": ""
                        }
                    }
                }
            },
            "pm_api_bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "pm_api_key": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization"
            }
        }
    },
    "security": [
        {
            "pm_api_auth_code": []
        },
        {
            "pm_api_bearer": []
        },
        {
            "pm_api_key": []
        }
    ]
}